feat(ui): move models to building blocks
