+++ apps/ui/src/__tests__/middleware.test.ts
@@ -43,6 +43,7 @@ describe("auth proxy", () => {
       "/evals/:path*",
       "/harnesses/:path*",
       "/mcp-servers/:path*",
+      "/models/:path*",
       "/orgs/:path*",
       "/sessions/:path*",
       "/settings/:path*",
