+++ apps/ui/src/__tests__/sidebar.test.tsx
@@ -164,6 +164,7 @@ describe("Sidebar", () => {
     expect(screen.getByText("Dashboard")).toBeInTheDocument();
     expect(screen.getByText("Harnesses")).toBeInTheDocument();
     expect(screen.getByText("Agents")).toBeInTheDocument();
+    expect(screen.getByText("Models")).toBeInTheDocument();
     expect(screen.getByText("Capabilities")).toBeInTheDocument();
     expect(screen.getByText("Settings")).toBeInTheDocument();
   });
@@ -174,12 +175,14 @@ describe("Sidebar", () => {
     const dashboardLink = screen.getByRole("link", { name: "Dashboard" });
     const harnessesLink = screen.getByRole("link", { name: "Harnesses" });
     const agentsLink = screen.getByRole("link", { name: "Agents" });
+    const modelsLink = screen.getByRole("link", { name: "Models" });
     const capabilitiesLink = screen.getByRole("link", { name: "Capabilities" });
     const settingsLink = screen.getByRole("link", { name: "Settings" });
 
     expect(dashboardLink).toHaveAttribute("href", "/dashboard");
     expect(harnessesLink).toHaveAttribute("href", "/harnesses");
     expect(agentsLink).toHaveAttribute("href", "/agents");
+    expect(modelsLink).toHaveAttribute("href", "/models");
     expect(capabilitiesLink).toHaveAttribute("href", "/capabilities");
     expect(settingsLink).toHaveAttribute("href", "/settings");
   });
@@ -231,7 +234,7 @@ describe("Sidebar", () => {
     expect(dashboardLink).toHaveClass("text-[13px]");
   });
 
-  it("has exactly 6 navigation items", () => {
+  it("has exactly 7 core navigation items", () => {
     render(<Sidebar />);
 
     // Get nav links (excluding logo link)
@@ -241,12 +244,19 @@ describe("Sidebar", () => {
         (link) =>
           link.getAttribute("href") !== "/dashboard" || link.textContent?.includes("Dashboard"),
       );
-    // Filter to only nav items (Chat, Dashboard, Harnesses, Agents, Capabilities, Settings)
-    const navItems = ["Chat", "Dashboard", "Harnesses", "Agents", "Capabilities", "Settings"];
+    const navItems = [
+      "Chat",
+      "Dashboard",
+      "Harnesses",
+      "Agents",
+      "Models",
+      "Capabilities",
+      "Settings",
+    ];
     const foundNavLinks = navLinks.filter((link) =>
       navItems.some((item) => link.textContent?.includes(item)),
     );
-    expect(foundNavLinks).toHaveLength(6);
+    expect(foundNavLinks).toHaveLength(7);
   });
 
   it("renders section labels for Building Blocks and Durable Execution", () => {
